feat: Handle wraparound and optimize space
We handle wrap around by making use of the generational counter idea and interpret the flag differently depending on the offset.

We also remove the need for client/server queue offsets as they are actually always the same value at all times, so we can make use of just one offset both in code and in the shm offset.

The impl is tested by making the test case use 256 RPC calls. This will wrap around, and the test case succeeds only if the impl was correct, which it is.