feat: Make RPC 0-copy
Fix is to perform the mmio read and write directly.
For result, this is trivial. We made an additional copy, for logging purposes to verify the result was correctly in shm. Now that we have verified it, we don't need to make this copy.

For the tuple, we can make use of a fold expression alongside the decay to write each