fix: Fixed two bugs with server-client flow and  mem wind
1. The first bug is a logic bug with the request channel

This was really a TODO that I left, that has now resurfaced as a bug. When we look for a server to service the request, the server can fail to set the memory window for whatever reason (say it has too many active clients) or (it had the 2nd bug, which is no longer a thing). Then, the FM should direct to another healthy server. This is done by now making the server respond with whether it has successfully spawned a server handler for the client. This involves a new IPC message type and a new handler in the bar 1 control write handler which simply redirects the response to the server.

2. The 2nd bug is with the channel closing. We were not properly erasing the offsets from the channel offsets map, so our MMIO offset was monotonically increasing up until the limit of the bar 2 window, and resulting in either a segfault, or a write that does not trigger the callback, thereby resulting in an infinite loop even tho both client and server are technically "connected"